+++ src/perks/constants/effects.tsx
