+++ src/perks/effects.tsx
@@ -1,4 +1,4 @@
-import * as constants from "../perks/constants/effects";
+import * as constants from "./constants/effects";
 
 export type IEffectStatistic =
   | "speed"
