+++ src/state/costs.test.tsx
@@ -1,0 +1,5 @@
+describe("Placeholder", () => {
+  test("should hold its place", () => {
+    return;
+  });
+});
